+++ bgstally/ui.py
@@ -45,9 +45,7 @@ class UI:
         self.image_button_cmdrs = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_cmdrs.png"))
         self.image_button_carrier = PhotoImage(file = path.join(self.bgstally.plugin_dir, FOLDER_ASSETS, "button_carrier.png"))
 
-        self.thread: Optional[Thread] = Thread(target=self._worker, name="BGSTally UI worker")
-        self.thread.daemon = True
-        self.thread.start()
+        self.indicate_activity:bool = False
 
         # Single-instance windows
         self.window_cmdrs:WindowCMDRs = WindowCMDRs(self.bgstally)
@@ -55,6 +53,10 @@ class UI:
         self.window_legend:WindowLegend = WindowLegend(self.bgstally)
         # TODO: When we support multiple APIs, this will no longer be a single instance window
         self.window_api:WindowAPI = WindowAPI(self.bgstally, self.bgstally.api_manager.apis[0])
+
+        self.thread: Optional[Thread] = Thread(target=self._worker, name="BGSTally UI worker")
+        self.thread.daemon = True
+        self.thread.start()
 
 
     def shut_down(self):
@@ -184,15 +186,33 @@ class UI:
                 Debug.logger.debug("Shutting down UI Worker...")
                 return
 
+            current_activity:Activity = self.bgstally.activity_manager.get_current_activity()
+
+            # Current Tick Time
             self.bgstally.overlay.display_message("tick", f"Curr Tick: {self.bgstally.tick.get_formatted(DATETIME_FORMAT_OVERLAY)}", True)
             minutes_delta:int = int((datetime.utcnow() - self.bgstally.tick.next_predicted()) / timedelta(minutes=1))
 
+            # Activity Indicator
+            if self.indicate_activity:
+                self.bgstally.overlay.display_indicator("indicator")
+                self.indicate_activity = False
+
+            # Tick Warning
             if datetime.utcnow() > self.bgstally.tick.next_predicted() + timedelta(minutes = TIME_TICK_ALERT_M):
                 self.bgstally.overlay.display_message("tickwarn", f"Tick {minutes_delta}m Overdue (Estimated)", True)
             elif datetime.utcnow() > self.bgstally.tick.next_predicted():
                 self.bgstally.overlay.display_message("tickwarn", f"Past Estimated Tick Time", True, text_colour_override="#FFA500")
             elif datetime.utcnow() > self.bgstally.tick.next_predicted() - timedelta(minutes = TIME_TICK_ALERT_M):
                 self.bgstally.overlay.display_message("tickwarn", f"Within {TIME_TICK_ALERT_M}m of Next Tick (Estimated)", True, text_colour_override="yellow")
+
+            # Thargoid War Progress Report
+            if (self.bgstally.state.system_tw_status is not None and current_activity is not None):
+                current_system:dict = current_activity.get_current_system()
+                if current_system:
+                    progress:float = float(self.bgstally.state.system_tw_status.get('WarProgress', 0))
+                    percent:float = round(progress * 100, 2)
+
+                    self.bgstally.overlay.display_progress_bar("tw", f"TW War Progress in {current_system.get('System', 'Unknown')}: {percent}%", progress)
 
             sleep(TIME_WORKER_PERIOD_S)